// frend/rb-tabulator/src/wc/services/pivot-types.ts
export type HeatmapMode = 'full' | 'row' | 'col';
// ...
export interface ChartJSData {
  labels: string[];
  datasets: ChartJSDataset[];
}

export interface ChartJSDataset {
  label: string;
  data: (number | null)[];
  backgroundColor?: string | string[];
  borderColor?: string | string[];
  type?: string;
}